split memory manager in separate file
